refactor: provide MaskRegistry as the central source of interfacing with mask data
