+++ junifer/data/masks/__init__.pyi
@@ -1,0 +1,6 @@
+__all__ = [
+    "MaskRegistry",
+    "compute_brain_mask",
+]
+
+from ._masks import MaskRegistry, compute_brain_mask
